refactor(tags): Cleanup and update remaining views and tests
