--- src/app/tags/components/TagForms/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./TagForms";
