--- src/app/tags/components/TagUpdate/TagUpdateFormFields/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./TagUpdateFormFields";
