--- src/app/tags/components/TagUpdate/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./TagUpdate";
