+++ src/app/tags/components/UpdateTagForm/UpdateTagFormFields/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./UpdateTagFormFields";
