+++ src/app/tags/components/UpdateTagForm/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./UpdateTagForm";
